@@ -2,7 +2,7 @@ package info.androidhive.fontawesome;
 
 import android.content.Context;
 import android.content.res.TypedArray;
-import android.support.v7.widget.AppCompatTextView;
+import androidx.appcompat.widget.AppCompatTextView;
 import android.util.AttributeSet;
 
 public class FontTextView extends AppCompatTextView {
